server/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic[email]==2.8.2
python-multipart==0.0.9
orjson==3.10.7
sqlmodel==0.0.16
psycopg[binary,pool]==3.2.3
pillow==11.0.0
